Moved filtering methods from ConversationBuilder to FilteredConversationBuilder
